Refractor Code for 1 API call and props
