+++ src/App.js
@@ -1,9 +1,14 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import axios from "axios";
 import "./App.css";
 import Title from "./Title";
 import Photo from "./Photo";
 import Caption from "./Caption";
 import styled from 'styled-components';
+
+const phDate = new Date();
+const dateFormat = phDate.getFullYear()+'-'+(phDate.getMonth()+1)+'-'+phDate.getDate();
+const URL = 'https://api.nasa.gov/planetary/apod?api_key=cpIVh1n5cT7gdxurNPJfIw7fYEakx1I89h1UEoYT';
 
 const Body = styled.div`
     background-color: #000;
@@ -21,6 +26,16 @@ padding: 2rem 0.1rem;
 `;
 
 function App() {
+  const [state,setState] = useState({});
+
+  useEffect( () => {
+    axios.get(`${URL}&date=${dateFormat}`)
+        .then( res => {
+            console.log(res.data);
+            setState(res.data)   
+        })
+        .catch(err => err);
+            },[])
   return (
     <Body className="App">
       <img className ='logo' src="https://upload.wikimedia.org/wikipedia/commons/e/e5/NASA_logo.svg" alt='nasa logo'></img>
@@ -28,9 +43,9 @@ function App() {
     '></img> */}
       <Header>Astronomy Photo of the Day</Header>
       
-      <Title />
-      <Photo />
-      <Caption />
+      <Title title={state.title} date = {state.date} />
+      <Photo photo = {state.url}/>
+      <Caption caption ={state.explanation} />
 
       {/* <p>
         Read through the instructions in the README.md file to build your NASA
