+++ src/Photo.js
@@ -1,10 +1,5 @@
 import React, { useEffect, useState } from "react";
-import axios from "axios";
 import styled from 'styled-components';
-
-const phDate = new Date();
-const dateFormat = phDate.getFullYear()+'-'+(phDate.getMonth()+1)+'-'+phDate.getDate();
-const URL = 'https://api.nasa.gov/planetary/apod?api_key=cpIVh1n5cT7gdxurNPJfIw7fYEakx1I89h1UEoYT';
 
 const Img = styled.img`
 width: 30rem;
@@ -14,19 +9,10 @@ object-fit: cover;
 margin: 3rem auto;
 `;
 
-const Photo = () => {
-    const [dailyPhoto , setDailyPhoto] = useState('');
-    useEffect( () => {
-        axios.get(`${URL}&date=${dateFormat}`)
-            .then( res => {
-                setDailyPhoto(res.data.url)
-                return dailyPhoto
-            })
-            .catch(err => err);
-                },[dailyPhoto])
+const Photo = (props) => {
                 return (
                 
-                    <Img src = {dailyPhoto} alt = 'daily astronomy'></Img>
+                    <Img src = {props.photo} alt = 'daily astronomy'></Img>
                 
                 
                 )
