+++ src/Title.js
@@ -1,11 +1,8 @@
 import React, { useEffect, useState } from "react";
-import axios from "axios";
 import styled from 'styled-components';
 
 
-const phDate = new Date();
-const dateFormat = phDate.getFullYear()+'-'+(phDate.getMonth()+1)+'-'+phDate.getDate();
-const URL = 'https://api.nasa.gov/planetary/apod?api_key=cpIVh1n5cT7gdxurNPJfIw7fYEakx1I89h1UEoYT';
+
 
 const APODTitle = styled.h2`
 font-size: 2rem;
@@ -15,20 +12,11 @@ padding-bottom: 1rem;
 
 `;
 
-const Title = () => {
-    const [photoTitle , setPhotoTitle] = useState('');
-    useEffect( () => {
-        axios.get(`${URL}&date=${dateFormat}`)
-            .then( res => {
-                setPhotoTitle(res.data.title)
-                return photoTitle
-            })
-            .catch(err => err);
-                },[photoTitle])
+const Title = (props) => {
                 return (
                 <div>
-                    <APODTitle>{photoTitle}</APODTitle>
-                    <p>{dateFormat}</p>
+                    <APODTitle>{props.title}</APODTitle>
+                    <p>{props.date}</p>
                 </div>
                 
                 )
